Reduce diffs when updating API docs (redux)
In a previous attempt (since reverted), instead of emitting links that contain the Git commit sha, emit the package's repo URL (with Git commit sha) once to a data file, and the use a shortcode in the generated docs to access the URL.

This ended up not working out very well as we already had a significant number of shortcodes on API docs pages (to enable markdown within HTML tags), and adding many more shortcodes to these pages caused Hugo to fallover trying to generate some of the larger API docs pages we have. To avoid the problem, we ended up reverting back to just including the Git commit sha within each link.

Since then some things have changed which should allow us to revisit this:

 1. By reorganizing how API docs pages are structured, we've since been able to eliminate the use of all shortcodes on API docs pages. The HTML is now structured so we don't need to use a shortcode to enable markdown inside HTML tags.

 2. We've increased Hugo's timeout for generating pages from the default of 10 seconds to 30 seconds, which should give Hugo some breathing room when generating large pages.

This change makes it so that instead of emitting links that contain the Git commit sha, we emit the Git commit sha in the page's front matter and then use Hugo's built-in `param` shortcode to lookup the value for each URL. So while this does introduce many shortcodes on API docs pages, this particular shortcode _should_ be much faster than the previous attempt that looked up the value in a data file and had some logic.

This should significantly reduce the size of diffs when updating API docs, making it easier to see/review the relevant changes.