Attempt to clarify open source vs. freemium vs. paid
This change:

- Add the Community Edition to the top-level edition tabs on
  the pricing page, replacing the separate section we already had.

- Adjust the feature differentiation waterfall in the edition tabs.

- Add a section to the top of the pricing page that clarifies that
  Pulumi is available as open source, freemium, or paid editions.

- Update the FAQ to call out the ability to use the open source
  independent of the service.

- Updates some text so that it doesn't sound like we're limiting
  the Community Edition to "individuals" -- in reality, it really
  just means "single user scenarios", which is subtly different.

This is certainly not perfect, but is a set of incremental changes
that I hope will clarify things at least a little bit.

The only thing I don't love is that it pushes the pricing tabs
closer to the fold, but I've adjusted some margins and kept the
new text fairly short, in an attempt to mitigate this.