Redirect GCP serverless page to cloudfunctions
The GCP provider no longer emits a `serverless` module as the APIs have been deprecated in favor of `cloudfunctions`. This has left behind a zombie docs page for the `serverless` module (since we do not delete files before regenerating docs).

Instead of keeping this file around, which will atrophy as we make format changes to doc pages, redirect to the `cloudfunctions` docs page.