Do an editorial pass over programming model docs
This is an editorial pass over the programming model docs, as part
of #1618. This includes rearranging ordering to minimize forward
references, getting rid of anything that smells "TypeScript-only",
restructuring headers so the ToC is a bit more structured,
wordsmithing, fleshing out examples, and trying to simplify the
presentation of content -- while also embellishing where it helps.

This is not done. The Go and TypeScript pages are empty. There is
still quite a number of TypeScript-only hyperlinks and terminology.