Use shortcodes to cut down on language-specific text/links
This change adds some shortcodes for common language-specific
text and associated links.

For instance, outputs are Output in JS, Output<T> in TS,
Output[T] in Python, Output in Go, and Output<T> in C#, and go
to very different places hyperlink-wise.

Because of the way our language switcher JavaScript works, we
display the JS version by default until it kicks in and picks
up the user's choice. This avoids awkward flashes in which all
or none of the options are available or missing.

I'm not done applying this to all possible codes, but at least
it lays the foundation to start paying down this debt.