Tweak PurgeCSS to include some missing CSS selectors
I noticed the CSS for styling a data source "symbol" was missing in the @pulumi/aws NodeJS API docs. The reason it's missing is because we only look or classes in the @pulumi/pulumi package, and that package doesn't have any datasources. To fix, just swap looking at @pulumi/aws rather than @pulumi/pulumi.