Fix Hugo `REF_NOT_FOUND` errors
We have a few relref links to "project", which currently work, but when
running the prototype API doc gen for AWS, it's generating project.md
files elsewhere in the site which causes Hugo to start failing for these
links with `REF_NOT_FOUND: Ref "project"`. This fixes these refs to point
to the right location of the project page.