Fix linter issues
The markdown linter that runs as part of `make build` was complaining about the following, which was failing the build:

```
Markdown Lint Results:
    - 427 files parsed.
    - 4 errors found.

Errors:

/Users/user/go/src/github.com/pulumi/docs/content/docs/get-started/install/_index.md:
Line 130: Fenced code blocks should be surrounded by blank lines.
Line 132: Fenced code blocks should be surrounded by blank lines.
Line 130: Lists should be surrounded by blank lines.
Line 131: Bare URL used.
``

This seems to be an issue with the linter, but in lieu of fixing the linter or disabling it around these lines, removing the indentation make it happy.