SEO optimization, KW "keda"
Primary KW: keda
Secondary KWs: kubernetes autoscaling, event driven autoscaling

Updated the metatags, the rest of the content is well optimized. Interestingly, "2020" came up in our semantic analysis, so if there is even a one-sentence update that can be made, that could be beneficial  — I'm thinking around "As of October 2019"...can we make this 2020? 

@zchase @cnunciato