Use next page title as the button in Getting Started
This changes the "NEXT STEP >" button to use the title of the
next page in the flow, like "INSTALL PULUMI >", etc. The theory
is that this is more enticing to click as well as clearly
communicating what comes next.

As part of this, I changed the "CONFIGURE >" step to be a little
clearer; for now, I chose "CONFIGURE CLOUD ACCESS >" for AWS,
Azure, and GCP, and "CONFIGURE CLUSTER ACCESS >" for Kubernetes.