add a repository dispatch event (#3522)
Our current docs build is run from within the provider repositories. Now
that we're moving to GitHub Actions we have an option to use a
repository dispatch event to trigger the builds from within this repo.

This event will not run unless a `POST` event is sent to this repo. It
will do the following:

- run typedocs
- run the resource generator docs
- run the python doc generator
- commit the changed files
- push them to a branch
- open a pull request

You can find an example of this in action here:

https://github.com/jaxxstorm/docs/pulls
https://github.com/jaxxstorm/docs/pull/2