add community packages page
This page is as a direct result of a conversation with @lukehoban and
the implementation/idea is up for debate.

Essentially, I've just added a table that allows us to link to well
known packages. Before I add a bunch of links, I'd like to get some
review on whether this is the right idea