Rethink provider docs generation github action (#5163)
Fixes: #5161

We are now in the situation where there are some providers that
require typedoc and python doc but not resource docs

This PR introduces the ability to be able to differentiate between
these style of providers as each of the providers will trigger a
specific event type

e.g. pulumi-aws will trigger tfgen-provider job
pulumi-awsx will trigger non-resource-provider job

the name `tfgen-provider` is a legacy name that was introduced mid-2020
but this will be replaced when all providers are updated. the logic
in the event.action allows that to be updated easily in the future by
adding a new event_type and phasing out the old one