Replace some readFileSyncs with FileAsset
The get-started tutorials for AWS and GCP still used the readFileSync
(or equivalent in other languages) way of reading contents for uploading
to cloud storage. There was also an instance in the Terraform migration
example.

The examples have all been changed to not use a subdirectory in order to
avoid the potential issue pointed out by Joe in #4903