unit testing guide: show how to set flag `dryRun` (#1642)
* Add PULUMI_TEST_MODE to guides/testing/unit

I was pulling my hair out trying to figure out why `pulumi.isDryRun`
was returning `false` when I was running my unit tests, given that
[the docs on the method][1] say:

> Note that we always consider executions in test mode to be
> “dry-runs”, since we will never actually carry out an update, and
> therefore certain output properties will never be resolved.

I’m still not sure if/where there is any canonical documentation on
“test mode” but I feel that it should probably always be enabled when
running unit tests. I’d prefer that there be a programmatic/imperative
way to enable it, e.g. `runtime.setDryRun(true)` or something similar,
but in lieu of that I feel the examples should include setting the env
var so as to plant seeds of awareness about this feature and how it
works.

[1]: https://www.pulumi.com/docs/reference/pkg/nodejs/pulumi/pulumi/runtime/#isDryRun

* Revert "Add PULUMI_TEST_MODE to guides/testing/unit"

f4adeec976811d64fe066b5084a9eeaf91b63814

As per discussion in pulumi/pulumi-hugo#1642

* guides/testing/unit: show how to set dryRun flag

Which, as per the discussion in pulumi/pulumi-hugo#1642, should
really/probably be renamed to `isPreview` or something along those
lines.

Anyway, it looks like the C# example already includes passing a value
for this flag, and I don’t know enough Go to add it to the Go example.

Now that I’m writing this commit message, the situation with `dryRun`
is confusing enough (especially given the docs on `runtime.isDryRun`)
that it could probably merit adding a small dedicated section to this
page. I’ll ask in the PR whether that’d be desirable.

* guides/testing/unit.md: better description of a flag

Co-authored-by: Ian Wahbe <ian@wahbe.com>

Co-authored-by: Ian Wahbe <ian@wahbe.com>