Fix Node.js SDK generation
The Pulumi Node.js SDK bumped the TypeScript version from 3.7.3 to 3.8.3
to be able to use OpenTelemetry. Unfortunately, this has broken the SDK
doc generation.

This change addresses the problem by upgrading TypeDoc to a version that
supports TypeScript 3.8.3 and fixing our tscdocgen tool to handle the
new output from the upgraded version of TypeDoc.