Fix generation of pulumi_policy python docs (#8224)
The problem was the doc generator was assuming that `__all__` was being used only for exporting submodules, which was the case for older generated python provider SDKs, but is not the case for `pulumi_policy`. `pulumi_policy` recently added `__all__` to its top-level `__init__.py` to explicitly define the top-level APIs exported by the package, none of which are modules. The doc generator was seeing `__all__` for `pulumi_policy` and was trying to use the `provider_with_module.rst` template, resulting in empty docs (it should be using the `provider_without_module.rst` template).

The fix is to be smarter about how we determine whether a module has submodules, but looking to see if the contents of `__all__` actually contains any modules.