set content-type for latest-dev-version explicitly
Set the content-type for the latest-dev-version to "text/plain"
explicitly.  The file will be uploaded anyway, but have the
"binary/octet-stream" content type.  This still works just fine, but
it looks better in a browser to have "text/plain", so the text will be
shown, instead of downloading a file.