fix: Automatically update Python docs (#10839)
When a new version of the CLI and Node.js and Python SDKs are released, we kick off a workflow to regenerate docs for these. However, the regenerated Python docs were not being included in the PR that is opened with the changes. This fixes that, by including the updated Python docs in the commit for the PR.