Python API docs: Avoid cross-referencing `pulumi_terraform` docs
I noticed when generating Python API reference docs locally, it was replacing references to `Any` in the main `pulumi` docs with a cross link to an `Any` type from `pulumi_terraform`, which is not what we want.

Using `:noindex:` appears to solve the problem. We don't list `pulumi_terraform` in the main index anyway--you can only get to it from the link to it on the Python language page. And, these `pulumi_terraform` docs should eventually go away and be moved into the registry anyway, if/when that package is schematized.