Add missing C# file content for new project
Fixes #12017

The bug was caused because not all language checkers had a `csharp` entry, and since the
language checker is global it doesn't make sense to display a language that can't be
selected for one checker.

Ideally, we would introduce a lint failure to make this mistake impossible (and catch
other places where we made the mistake). That is out of scope for this PR.