Experiment with different docs nav structure
This splits out IaC, ESC, and Cloud into their own independent
top level nav structures. This simplifies and focuses the left nav
for each one of these. There are some gray areas, like conceptual
topics in IaC (or Cloud) that do apply to one or the other, but I
personally find this to be way more usable than the current mammoth
combined left nav. As you go between the products, the left nav
drops the "Pulumi IaC", "Pulumi ESC", and "Pulumi Cloud" headers in
the nav since they're superfluous given the top level tabs.

This is not final and we've certainly not built consensus as a team
on this yet -- just pushing the branch so we have something to discuss!