automatically approve PRs that were created by pulumi-bot
pulumi-bot creates documentation update and version update PRs in this
repo.  These should be auto-merged, but that's broken because we've
set up branch protections, and the action we're using to auto-merge
PRs won't go ahead and try to merge PRs if they are not approved (even
if there are exceptions set up in the branch protections to allow
these PRs through).

Auto-approve PRs that come from pulumi-bot.  This should make the
auto-merge automation work again.