Fix a few rounded buttons (#12895)
We missed a handful of round buttons in the homepage update, which
this change addresses. We also omitted one of the styling changes in
the original PR which was to make the CTA buttons less cartoonishly
large. Wasn't sure if that was intentional, but this adds it back.

Fixes pulumi/docs#12793.