Fix incorrect formatting of note due to significant whitespace
The 6 space indent on the notes block here is causing it to be interpreted as a pre-formatted code block. Removing the indentation gives it the correct formatting and it's still included in the "Audience" bullet point in terms of nesting within that structure, as was intended by the original author.