Document Python output lifting's use of `__getattr__` (#13103)
The Pulumi Python SDK lifts output properties by overriding the special
`__getattr__` method. This is designed to improve the ergonomics of
writing Pulumi programs in Python, but technically violates some of the
contracts Python expects of this method, particularly in relation to
`hasattr`. This commit warns of this potential gotcha when mentioning
output lifting to give users a reference point should they bump into
this issue.

Fixes pulumi/pulumi#16743