Rewrite how Pulumi works page to use BucketV2 (#13142)
Bucket is going to be deprecated and BucketV2 is recommended. IN addition to that, bucket.acl is deprecated. Unfortunately these resources were selected to demonstrate how Pulumi works. The page itself does not need to concern itself with Bucket per se, but needs a good example. To avoid the difficulty with deprecated resources and attributes, this rewrite of how-pulumi-works uses BucketV2 and tags modification as the example, instead of using Bucket and acl.

Relates: pulumi/home#3583