Update supported version of macOS
The docs are referring to a _very_ old version of macOS. Apple isn't forthcoming with official supported dates, but Apple usually provides security updates for the latest 3 releases only. The oldest supported version we can use to test on GitHub Actions is macOS 13, so this is the version we officially support. Update the docs to reflect this.