Upgrade markdown-lint to v0.37.4 (#14015)
* Upgrade markdown-lint to v0.37.4

This improves the linter, and catches multiple instances of invalid markdown that were
shown to customers.

* Normalize the environment for the lint-markdown script

By using `yarn run` we get to automatically tap into the dependency tree of the
package. By using `ts-node` and a typescript file, we can opt into better typing. This
already discovered a repeated literal in an object.

* Make the module decision explicit

This is necessary to include modern dependencies like "markdownlint".

* Replace `ts-node` with `tsx`

* Don't invoke empty cypress plugin

* Move /scripts/search to TS