Skip failing docs examples (#14506)
There are about 20 failing example programs in master; this results in an alert in the #docs-ops channel everyday and makes it hard for us to notice and fix any new failures in the 300+ programs that are currently working!

This PR disables those failing test so we can catch and fix new failures in the future. I've also filed #14505 to track fixing these so we can ratchet this up to covering all of these programs.