Blog about running the program for `refresh` and `destroy`
As of the latest release of Pulumi, `pulumi refresh` and `pulumi
destroy` support a new `--run-program` flag, which instructs Pulumi to
run the user's program before performing an operation. This is in
contrast to the default behaviour, where Pulumi will simply operate on
existing state, which can cause issues when the program's execution is
required to e.g. produce new short-lived credentials.