Blog about running the program for `refresh` and `destroy` (#14627)
* Blog about running the program for `refresh` and `destroy`

As of the latest release of Pulumi, `pulumi refresh` and `pulumi
destroy` support a new `--run-program` flag, which instructs Pulumi to
run the user's program before performing an operation. This is in
contrast to the default behaviour, where Pulumi will simply operate on
existing state, which can cause issues when the program's execution is
required to e.g. produce new short-lived credentials.

* minor tweeks and fleshed out uses cases

* Update index.md

* Reworked the use cases list

* Update index.md

* Update index.md

* Update content/blog/improved-refresh-destroy-experience/index.md

Co-authored-by: Fraser Waters <fraser@pulumi.com>

* fixing based on Fraser's feedback

* Finishing touches and updates

* Update index.md

---------

Co-authored-by: meagancojocar <50200557+meagancojocar@users.noreply.github.com>
Co-authored-by: Fraser Waters <fraser@pulumi.com>