Don't imply the CLI isn't required for Auto API (#15917)
The intent behind "without requiring invoking the CLI from a shell process" is to highlight that you don't need to run the `pulumi` binary "by hand", but it might lead people to think that you don't need the `pulumi` binary. This change rewords the sentence so it doesn't imply the `pulumi` binary isn't required.