Update pulumi-command shortcode to use local file (#16072)
The `pulumi-command` shortcode is used to emit an inline copy of the CLI's root command docs, which is used on a couple pages:

- https://www.pulumi.com/docs/iac/cli/
- https://www.pulumi.com/docs/iac/cli/commands/

Originally, this shortcode simply opened the local file. Then when the docs were split between pulumi/docs and pulumi/pulumi-hugo repos, the shortcode was changed to make an API request to GitHub to grab the file's content from the docs repo.

We've since merged these two repos back into the single pulumi/docs repo, so there's no longer a need to make a network request to get the content. We can go back to just reading the file. Which is what this change does.

This should make site generation slightly faster and less flaky.