esc: move syntax reference under environments
These changes move the ESC Environment syntax reference from the
reference portion of the docs to the Environments portion of the docs.

As it stands, the syntax reference--which is a primary source of
information for Environment authors--is not particularly discoverable,
as it is completely disconnected from the rest of the ESC documentation.