Document provider layering, with impl guides
Today we do not really explain the layering of our various means of
writing providers, which as we've seen, can lead to confusion.

This does a pass over the provider authoring docs, explaining the
architecture, and why only Go has a higher level SDK.

It also adds reference material and implementation guides for those who
want to code to the direct gRPC/Protobuf interfaces directly, in Python,
Go, or TypeScript. (These were tested locally.)

I am not 100% convinced we should add this but for power users wanting
to really understand how providers work and how to author them, it would
seem helpful in demystifying some confusing aspects of our ecosystem.